fix warnings

+++ ya/interview/aa1/oneEditApart/main.cpp
@@ -6,7 +6,7 @@
 bool oneEditApart(const std::string& first, const std::string& second) {
     const std::string& small = first.size() <= second.size() ? first : second;
     const std::string& big = first.size() > second.size() ? first : second;
-    int lenDiff = big.size() - small.size();
+    int lenDiff = static_cast<int>(big.size() - small.size());
     if (lenDiff > 1) {
         return false;
     }
